crypto: refactor and reorganize crypto.js
