+++ node.gyp
@@ -87,6 +87,14 @@
       'lib/internal/v8_prof_polyfill.js',
       'lib/internal/v8_prof_processor.js',
       'lib/internal/streams/lazy_transform.js',
+      'lib/internal/crypto/certificate.js',
+      'lib/internal/crypto/cipherbase.js',
+      'lib/internal/crypto/dhbase.js',
+      'lib/internal/crypto/hashbase.js',
+      'lib/internal/crypto/pbkdf2.js',
+      'lib/internal/crypto/rsa.js',
+      'lib/internal/crypto/signbase.js',
+      'lib/internal/crypto/util.js',
       'deps/v8/tools/splaytree.js',
       'deps/v8/tools/codemap.js',
       'deps/v8/tools/consarray.js',
